Исправление ошибок на null или пустоту выборки

+++ idcc/Repository/Interfaces/IDataRepository.cs
@@ -10,7 +10,7 @@ public interface IDataRepository
     Task<(double, Grade)> GetGradeLevelAsync(double score);
 
     Task<double> GetPercentOrDefaultAsync(string code, double value);
-    Task<double> GetCountOrDefaultAsync(string code, double value);
+    Task<int> GetCountOrDefaultAsync(string code, int value);
 
     Task<(Grade? prev, Grade? next)> GetRelationAsync(Grade current);
 }
